+++ sia-backend/.env
@@ -1,2 +1,7 @@
 DATABASE_PASSWORD="postgres"
-SECRET_KEY="b6w9DHbVksLaEQLU4!^q@8xXE$&6YYDjibSDEkK5tU&x#c#Z2v"
+DATABASE_HOST="database"
+DATABASE_PORT="5432"
+DATABASE_NAME="postgres"
+DATABASE_USER="postgres"
+DEFAULT_USER="owner"
+DEFAULT_USER_PASSWORD="562d5ea007120c4c90ff3c4c114fc1affe3315d14220b7357f"
