+++ sia-backend/flask-server/config.py
@@ -1,0 +1,7 @@
+import os
+import secrets
+
+class Config:
+    SQLALCHEMY_DATABASE_URI = os.getenv("DATABASE_URI")
+    DEBUG = os.getenv("DEBUG")
+    SECRET_KEY = secrets.token_hex(25)
