+++ sia-backend/flask-server/static/css/base.css
@@ -1,5 +1,4 @@
 .class_footer{
     width: 100%;
-
     background-color: rgb(43, 48, 53);
 }
